[C8-17] Improve portability of JVM source code when using gcc7.
This patch places up-to-date type declarations in all places where the gcc switch “-Wno-deprecated-declarations” would flag problems and it enables the switch to catch future related issues. This makes the source code compile on all present Amazon Linux versions.

+++ src/jdk/src/share/native/sun/font/freetypeScaler.c
@@ -1465,7 +1465,7 @@ Java_sun_font_FreetypeFontScaler_getGlyphVectorOutlineNative(
              (FTScalerInfo*) jlong_to_ptr(pScaler);
 
     glyphs = NULL;
-    if (numGlyphs > 0 && 0xffffffffu / sizeof(jint) >= numGlyphs) {
+    if (numGlyphs > 0 && 0xffffffffu / sizeof(jint) >= (size_t)numGlyphs) {
         glyphs = (jint*) malloc(numGlyphs*sizeof(jint));
     }
     if (glyphs == NULL) {
